+++ lib/legion/telemetry.rb
@@ -2,6 +2,9 @@
 
 module Legion
   module Telemetry
+    autoload :OpenInference, 'legion/telemetry/open_inference'
+    autoload :SafetyMetrics, 'legion/telemetry/safety_metrics'
+
     module_function
 
     def otel_available?